+++ src/bin/lc.rs
@@ -1,8 +1,14 @@
 use leetcode_cli::cli;
+use tokio::runtime::Builder;
 
 fn main() {
-    let r = cli::main();
-    if r.is_err() {
-        println!("{:?}", r.err().expect("This won't happend."));
+    if let Err(err) = Builder::new()
+        .basic_scheduler()
+        .enable_all()
+        .build()
+        .expect("Build tokio runtime failed")
+        .block_on(cli::main())
+    {
+        println!("{:?}", err);
     }
 }